Improve Map
- Level, 0-1 맵 수정
- 플레이어 킬 오류 수정
- 플레이어 1인칭 카메라 복구

+++ Mafia/Assets/Charact_Map/Character/Scripts/PlayerController.cs
@@ -114,7 +114,7 @@ public class PlayerController : MonoBehaviourPun
             if (Input.GetMouseButtonDown(0))
             {
                 Debug.Log("Before Kill()");
-                if (Time.time - lastKillTime >= killCooldown)
+                if (Time.time - lastKillTime >= killCooldown && PhotonNetwork.LocalPlayer.CustomProperties.ContainsKey("isMafia"))
                 {
                     playerAnimator.Kill();
                     lastKillTime = Time.time;
